requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
boto3==1.16.25
botocore==1.19.25
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptg==0.2.post2
dnspython==1.16.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
h11==0.11.0
h2==4.0.0
hachoir==3.1.1
hpack==4.0.0
httplib2==0.18.1
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.10
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
multidict==5.0.2
oauthlib==3.1.0
parso==0.7.1
Pillow==8.0.1
priority==1.3.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptpython==3.0.7
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
python-dateutil==2.8.1
pytz==2020.4
Quart==0.14.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
Telethon==1.18.2
TgCrypto==1.2.2
toml==0.10.2
tqdm==4.54.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wsproto==1.0.0
yarl==1.6.3
